+++ data/store/src/backend/mock/store.rs
@@ -9,6 +9,10 @@ use replicante_data_models::Event;
 use replicante_data_models::Node;
 use replicante_data_models::Shard;
 
+use super::super::super::EventsFilters;
+use super::super::super::EventsIter;
+use super::super::super::EventsOptions;
+
 use super::super::super::Result;
 use super::super::super::store::InnerStore;
 
@@ -48,6 +52,13 @@ impl InnerStore for MockStore {
         let clusters = self.clusters_meta.lock().unwrap();
         let meta = clusters.get(&cluster).cloned();
         Ok(meta)
+    }
+
+    fn events(&self, _filters: EventsFilters, _options: EventsOptions) -> Result<EventsIter> {
+        let events = self.events.lock().unwrap().clone();
+        let events: Vec<_> = events.into_iter().rev().collect();
+        let iter = events.into_iter().map(|e| Ok(e));
+        Ok(EventsIter::new(iter))
     }
 
     fn find_clusters(&self, search: String, _: u8) -> Result<Vec<ClusterMeta>> {