+++ data/store/src/lib.rs
@@ -33,7 +33,12 @@ pub use self::backend::mock;
 
 pub use self::config::Config;
 pub use self::errors::*;
+
+pub use self::store::EventsFilters;
+pub use self::store::EventsIter;
+pub use self::store::EventsOptions;
 pub use self::store::Store;
+
 pub use self::validator::ValidationResult;
 pub use self::validator::Validator;