+++ data/store/src/store.rs
@@ -16,6 +16,61 @@ use super::Result;
 use super::backend::mongo::MongoStore;
 
 
+/// Iterator over events returned by `Store::events`.
+pub struct EventsIter(Box<Iterator<Item=Result<Event>>>);
+
+impl EventsIter {
+    pub fn new<I>(iter: I) -> EventsIter
+        where I: Iterator<Item=Result<Event>> + 'static
+    {
+        EventsIter(Box::new(iter))
+    }
+}
+
+impl Iterator for EventsIter {
+    type Item = Result<Event>;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next()
+    }
+}
+
+
+/// Filters to apply when iterating over events.
+pub struct EventsFilters {}
+
+impl EventsFilters {
+    /// Return all events, don't skip any.
+    pub fn all() -> EventsFilters {
+        Self::default()
+    }
+}
+
+impl Default for EventsFilters {
+    fn default() -> EventsFilters {
+        EventsFilters { }
+    }
+}
+
+
+/// Options to apply when iterating over events.
+pub struct EventsOptions {
+    /// Max number of events to return.
+    pub limit: Option<i64>,
+
+    /// By default events are returned old to new, set to true to reverse the order.
+    pub reverse: bool,
+}
+
+impl Default for EventsOptions {
+    fn default() -> EventsOptions {
+        EventsOptions {
+            limit: None,
+            reverse: false,
+        }
+    }
+}
+
+
 /// Private interface to the persistence storage layer.
 ///
 /// Allows multiple possible datastores to be used as well as mocks for testing.
@@ -31,6 +86,9 @@ pub trait InnerStore: Send + Sync {
 
     /// See `Store::cluster_meta` for details.
     fn cluster_meta(&self, cluster: String) -> Result<Option<ClusterMeta>>;
+
+    /// See `Store::events` for details.
+    fn events(&self, filters: EventsFilters, options: EventsOptions) -> Result<EventsIter>;
 
     /// See `Store::find_clusters` for details.
     fn find_clusters(&self, search: String, limit: u8) -> Result<Vec<ClusterMeta>>;
@@ -137,6 +195,14 @@ impl Store {
         where S: Into<String>,
     {
         self.0.cluster_meta(cluster.into())
+    }
+
+    /// Return an iterator over events in the store.
+    ///
+    /// Pass `filters` to tune the events that will be returned and `options` to
+    /// control result behavior like limit of items or order (old to new/new to old).
+    pub fn events(&self, filters: EventsFilters, options: EventsOptions) -> Result<EventsIter> {
+        self.0.events(filters, options)
     }
 
     /// Search for a list of clusters with names matching the search term.