+++ streams/events/src/mock.rs
@@ -3,7 +3,10 @@ use std::sync::Arc;
 use replicante_data_models::Event;
 
 use super::EventsStream;
+use super::Iter;
 use super::Result;
+use super::ScanFilters;
+use super::ScanOptions;
 use super::interface::StreamInterface;
 
 
@@ -22,6 +25,10 @@ impl MockEvents {
 
 impl StreamInterface for MockEvents {
     fn emit(&self, _event: Event) -> Result<()> {
+        Err("Not yet implemented".into())
+    }
+
+    fn scan(&self, _filters: ScanFilters, _options: ScanOptions) -> Result<Iter> {
         Err("Not yet implemented".into())
     }
 }